[MODEL] Fixed a problem where `Hashie::Mash#min` and `#max` returned unexpected values
While people can define names for aggregations such as `min` and `max`, these
methods are defined in `Enumerable#min` and `Enumerable#max`, and return
an unexpected value:

    result.aggregations
    # => {"price"=>{"doc_count"=>9428, "min"=>{"value"=>1.0}, "max"=>{"value"=>260000.0}}}
    result.aggregations.price
    # => {"doc_count"=>9428, "min"=>{"value"=>1.0}, "max"=>{"value"=>260000.0}}
    result.aggregations.price.min
    # => ["doc_count", 9428]

Therefore, the whole Hash with aggregations is being traversed, in a wrapper class,
and a `min` and `max` method is defined for Hashie::Mash instances,
because these names are too common.

The wrapper class can have its uses further down the line, so a dedicated
unit test file has been added as well.

Closes #568

Related: elastic/elasticsearch-ruby#306